+++ src/plugins/plugins.ts
@@ -1,28 +1,25 @@
 import { backgroundConfigs } from "./backgrounds";
 import { widgetConfigs } from "./widgets";
-import React from "react";
-
 export { backgroundConfigs } from "./backgrounds";
 export { widgetConfigs } from "./widgets";
 
-const configs = [...backgroundConfigs, ...widgetConfigs];
+import Unknown from "./widgets/unknown/Unknown";
+import UnknownSettings from "./widgets/unknown/UnknownSettings";
 
-const DefaultComponent: React.FC = () => "Component not available";
+const configs = [...backgroundConfigs, ...widgetConfigs];
 
 export function getConfig(key: string) {
   const config = configs.find((config) => config.key === key);
 
-  // if (!config) throw new Error(`Unable to find config for plugin: ${key}`); // WHY WOULD YOU BREAK THE ENTIRE APP IF A PLUGIN IS MISSING
   if (!config) {
-    console.error(`Unable to find config for plugin: ${key}`);
+    console.warn(`Unable to find config for plugin: ${key}`);
     return {
-      key,
-      name: key,
-      description:
-        "The plugin id is not valid, if you need help, create an issue on the github repo.",
-      settingsComponent: DefaultComponent,
+      key: "widget/unknown",
+      name: `Unknown Widget ${key}`,
+      description: `Something went wrong while loading '${key}' (click for details)`,
+      dashboardComponent: Unknown,
+      settingsComponent: UnknownSettings,
       supportsBackdrop: false,
-      dashboardComponent: DefaultComponent,
     };
   }